+++ relay-server/src/processing/trace_metrics/process.rs
@@ -107,6 +107,9 @@ fn normalize_trace_metric(
 
     if let Some(metric_value) = metric.value_mut() {
         eap::trace_metric::normalize_metric_name(metric_value)?;
+        if ctx.is_processing() {
+            eap::trace_metric::normalize_metric_unit(metric_value);
+        }
         eap::normalize_received(&mut metric_value.attributes, meta.received_at());
         eap::normalize_client_address(&mut metric_value.attributes, meta.client_addr());
         eap::normalize_user_agent(